requirements.txt
pytest==8.3.5
selenium==4.32.0
webdriver-manager==4.0.2
python-dotenv==1.1.0
ruff==0.11.8
allure-pytest==2.14.2 
allure-python-commons==2.14.2